// src/model/tool.ts
import { z } from "zod";
import { Context } from "./context";
import { Session } from "./session";

export interface Tool<ISchema extends z.ZodType, O> {
  name: string;
  description: string;
  inputSchema: ISchema;
  // input is the parameters for the tool, which is expected to be generated by the AI agent
  handler: (context: Context, session: Session, input: z.infer<ISchema>) => Promise<O>;
}